refactor: move assignRawDocsToIdStructure to its own file
