+++ lib/model.js
@@ -23,6 +23,7 @@ const applyHooks = require('./helpers/model/applyHooks');
 const applyMethods = require('./helpers/model/applyMethods');
 const applyStatics = require('./helpers/model/applyStatics');
 const applyWriteConcern = require('./helpers/schema/applyWriteConcern');
+const assignRawDocsToIdStructure = require('./helpers/populate/assignRawDocsToIdStructure');
 const cast = require('./cast');
 const castUpdate = require('./helpers/query/castUpdate');
 const discriminator = require('./helpers/model/discriminator');
@@ -3614,7 +3615,9 @@ function populate(model, docs, options, callback) {
   }
 
   function _assign(err, vals, mod, assignmentOpts) {
-    if (err) return callback(err, null);
+    if (err) {
+      return callback(err, null);
+    }
 
     const options = mod.options;
     const isVirtual = mod.isVirtual;
@@ -3724,8 +3727,7 @@ function assignVals(o) {
 
   // replace the original ids in our intermediate _ids structure
   // with the documents found by query
-  assignRawDocsToIdStructure(o.rawIds, o.rawDocs, o.rawOrder, populateOptions,
-    o.localField, o.foreignField);
+  assignRawDocsToIdStructure(o.rawIds, o.rawDocs, o.rawOrder, populateOptions);
 
   // now update the original documents being populated using the
   // result structure that contains real documents.
@@ -3767,124 +3769,45 @@ function assignVals(o) {
       // If virtual populate and doc is already init-ed, need to walk through
       // the actual doc to set rather than setting `_doc` directly
       mpath.set(o.path, rawIds[i], docs[i], setValue);
-    } else {
-      const parts = o.path.split('.');
-      let cur = docs[i];
-      for (let j = 0; j < parts.length - 1; ++j) {
-        if (cur[parts[j]] == null) {
-          cur[parts[j]] = {};
-        }
-        cur = cur[parts[j]];
-      }
-      if (docs[i].$__) {
-        docs[i].populated(o.path, o.allIds[i], o.allOptions);
-      }
+      continue;
+    }
 
-      if (o.isVirtual && o.lean) {
-        // If lean, need to check that each individual virtual respects
-        // `justOne`, because you may have a populated virtual with `justOne`
-        // underneath an array. See gh-6867
-        utils.setValue(o.path, rawIds[i], docs[i], function(v) {
-          if (o.justOne && Array.isArray(v)) {
-            return setValue(v[0]);
-          } else if (!o.justOne && !Array.isArray(v)) {
-            return setValue([v]);
-          }
-          return setValue(v);
-        }, false);
-      } else {
-        utils.setValue(o.path, rawIds[i], docs[i], setValue, false);
+    const parts = o.path.split('.');
+    let cur = docs[i];
+    for (let j = 0; j < parts.length - 1; ++j) {
+      if (cur[parts[j]] == null) {
+        cur[parts[j]] = {};
       }
+      cur = cur[parts[j]];
+    }
+    if (docs[i].$__) {
+      docs[i].populated(o.path, o.allIds[i], o.allOptions);
+    }
+
+    if (o.isVirtual && o.lean) {
+      // If lean, need to check that each individual virtual respects
+      // `justOne`, because you may have a populated virtual with `justOne`
+      // underneath an array. See gh-6867
+      utils.setValue(o.path, rawIds[i], docs[i], function(v) {
+        if (o.justOne && Array.isArray(v)) {
+          return setValue(v[0]);
+        } else if (!o.justOne && !Array.isArray(v)) {
+          return setValue([v]);
+        }
+        return setValue(v);
+      }, false);
+    } else {
+      utils.setValue(o.path, rawIds[i], docs[i], setValue, false);
     }
   }
 }
 
 /*!
- *
+ * Check if obj is a document
  */
 
 function isModel(obj) {
   return get(obj, '$__') != null;
-}
-
-/*!
- * Assign `vals` returned by mongo query to the `rawIds`
- * structure returned from utils.getVals() honoring
- * query sort order if specified by user.
- *
- * This can be optimized.
- *
- * Rules:
- *
- *   if the value of the path is not an array, use findOne rules, else find.
- *   for findOne the results are assigned directly to doc path (including null results).
- *   for find, if user specified sort order, results are assigned directly
- *   else documents are put back in original order of array if found in results
- *
- * @param {Array} rawIds
- * @param {Array} vals
- * @param {Boolean} sort
- * @api private
- */
-
-function assignRawDocsToIdStructure(rawIds, resultDocs, resultOrder, options, localFields, foreignFields, recursed) {
-  // honor user specified sort order
-  const newOrder = [];
-  const sorting = options.sort && rawIds.length > 1;
-  let doc;
-  let sid;
-  let id;
-
-  for (let i = 0; i < rawIds.length; ++i) {
-    id = rawIds[i];
-
-    if (Array.isArray(id)) {
-      // handle [ [id0, id2], [id3] ]
-      assignRawDocsToIdStructure(id, resultDocs, resultOrder, options, localFields, foreignFields, true);
-      newOrder.push(id);
-      continue;
-    }
-
-    if (id === null && !sorting) {
-      // keep nulls for findOne unless sorting, which always
-      // removes them (backward compat)
-      newOrder.push(id);
-      continue;
-    }
-
-    sid = String(id);
-
-    if (recursed) {
-      // apply find behavior
-
-      // assign matching documents in original order unless sorting
-      doc = resultDocs[sid];
-      if (doc) {
-        if (sorting) {
-          newOrder[resultOrder[sid]] = doc;
-        } else {
-          newOrder.push(doc);
-        }
-      } else {
-        newOrder.push(id);
-      }
-    } else {
-      // apply findOne behavior - if document in results, assign, else assign null
-      newOrder[i] = doc = resultDocs[sid] || null;
-    }
-  }
-
-  rawIds.length = 0;
-  if (newOrder.length) {
-    // reassign the documents based on corrected order
-
-    // forEach skips over sparse entries in arrays so we
-    // can safely use this to our advantage dealing with sorted
-    // result sets too.
-    newOrder.forEach(function(doc, i) {
-      rawIds[i] = doc;
-    });
-  }
 }
 
 function getModelsMapForPopulate(model, docs, options) {
